requirements.txt
asgiref==3.3.1
boto3==1.16.25
botocore==1.19.25
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Django==3.1.3
django-rest-knox==4.1.0
django-storages==1.10.1
djangorestframework==3.12.2
Flask==1.1.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
Werkzeug==1.0.1
